WIP now

+++ src/main/java/com/evoter/vote/dto/AddVoteRequest.java
@@ -2,7 +2,7 @@ package com.evoter.vote.dto;
 
 public record AddVoteRequest(
         Long userId,
-        Integer candidateId,
+        Long candidateId,
         Integer pollId
 ) {
 }
